+++ Validation/PhysicsTools/src/PATValidation_MET.cc
@@ -13,7 +13,7 @@
 //
 // Original Author:  Sudhir_Malik
 //         Created:  Fri Mar 13 09:52:17 CDT 2009
-// $Id: PATValidation_MET.cc,v 1.5 2010/02/17 09:30:38 kfjack Exp $
+// $Id: PATValidation_MET.cc,v 1.6 2010/03/25 16:30:23 kfjack Exp $
 //
 //
 
@@ -181,7 +181,7 @@ PATValidation_MET::analyze(const edm::Event& iEvent, const edm::EventSetup& iSet
      iEvent.getByLabel(recoMET_, recomet_hnd);
      const candidateCollection & RECOMET =*recomet_hnd;
     // Loop over Reco Mets
-     for (int i = 0; i != RECOMET.size(); ++i) {
+     for (int i = 0; i != (int)RECOMET.size(); ++i) {
       me["RecoMet_pt"]->Fill(RECOMET[i].pt());
 	 nRecoMET++;
 //     cout << "RecoMET pt is = " << RECOMET[i].pt() << endl;
@@ -214,7 +214,7 @@ PATValidation_MET::analyze(const edm::Event& iEvent, const edm::EventSetup& iSet
      iEvent.getByLabel(patMET_, patmet_hnd);
      std::vector<pat::MET> const & PATMET = *patmet_hnd;
    // Loop over Pat Mets
-     for (int i = 0; i != PATMET.size(); ++i) {
+     for (int i = 0; i != (int)PATMET.size(); ++i) {
      me["PatMet_pt"]->Fill(PATMET[i].pt());
      nPatMET++;
   //      cout << "PatMET pt is " << PATMET[i].pt() << endl;
